Added tile entity asset managers, some testing for chunk backgrounds, added square tile borders

+++ Assets/Scripts/World/Creation/ProcGenHelper.cs
@@ -57,7 +57,7 @@ namespace WorldModule {
 
             SeralizedFluidTileData seralizedFluidTileData = new SeralizedFluidTileData();
             seralizedFluidTileData.ids = new string[size.x,size.y];
-            seralizedFluidTileData.fill = new int[size.x,size.y];
+            seralizedFluidTileData.fill = new float[size.x,size.y];
 
             SeralizedChunkConduitData itemConduitData = new SeralizedChunkConduitData();
             itemConduitData.ids = new string[size.x,size.y];
@@ -152,7 +152,7 @@ namespace WorldModule {
 
             SeralizedFluidTileData fluidData = new SeralizedFluidTileData();
             fluidData.ids = new string[Global.ChunkPartitionSize,Global.ChunkPartitionSize];
-            fluidData.fill = new int[Global.ChunkPartitionSize,Global.ChunkPartitionSize];
+            fluidData.fill = new float[Global.ChunkPartitionSize,Global.ChunkPartitionSize];
 
             for (int tileX = 0; tileX < Global.ChunkPartitionSize; tileX ++) {
                 for (int tileY = 0; tileY < Global.ChunkPartitionSize; tileY ++) {
